Add an unfinished Game Boy emulator
For the moment it only implements part of the CPU instruction set and
very basic memory management.

Graphics, audio and interrupt handling are completely missing.

I've implemented just barely enough to be able to run the instructions
from the Game Boy's boot ROM. Since I did not implement the PPU yet,
I've had to hardcode the LCDC Y-Coordinate (0xFF44) to 0x90 so that I
didn't end up in an endless loop.